✨ Add loading scene with match details
