+++ game/index.html
@@ -8,5 +8,6 @@
   </head>
   <body>
     <script src="/game.js" type="module"></script>
+    <div id="game"></div>
   </body>
 </html>
